Update the PowerSelect stories
- extract the PowerSelectMultiple example into a dedicated file
- add a disabled state
